refactor: removed unnecessary folder structure
